Extract LoadOrganizations to new file
