+++ website/src/index.tsx
@@ -4,8 +4,9 @@ import * as React from "react";
 import { render } from "react-dom";
 import { Provider } from "react-redux";
 import { createStore } from "redux";
-import { SetOrganizations} from "./Actions";
+import { SetOrganizations } from "./Actions";
 import { App } from "./App";
+import { LoadOrganizations } from "./DataLoader";
 import { RootReducer } from "./Reducers/RootReducer";
 import { County } from "./State/County";
 import { Gender } from "./State/Gender";
@@ -27,45 +28,14 @@ const store = createStore(
   ),
 );
 
-function loadJson<T>(url: string): Promise<T>
-{
-  return fetch(url)
-    .then((response) =>
-    {
-      if (!response.ok)
-      {
-        throw new Error(response.statusText);
-      }
-      return response.json() as Promise<T>;
-    });
-}
-
-function parseOrganization(data: any): Organization
-{
-  return {
-    ...data,
-    countiesServed: Set(data.countiesServed),
-    servicesCategories: Set(data.servicesCategories),
-  };
-}
-
-export function LoadOrganizations(): Promise<Set<Organization>>
-{
-  return loadJson<any[]>("./data.json").then((data) =>
-  {
-    return Set<Organization>(data.map(parseOrganization));
-  });
-}
-
-LoadOrganizations().then((organizations) =>
-{
+LoadOrganizations().then((organizations) => {
   store.dispatch(SetOrganizations(organizations));
 });
 
 render(
   (
     <Provider store={store}>
-      <App/>
+      <App />
     </Provider>
   ),
   document.getElementById("root") as HTMLElement,
